requirements.txt
pystray
notify-py
requests
pystray
misskey.py